Refactor git module
