+++ src/lib.rs
@@ -1,9 +1,9 @@
 pub mod config;
+mod git;
 
 use anyhow::{anyhow, Error, Result};
-use chrono::{TimeZone, Utc};
+use chrono::Utc;
 use config::Config;
-use git2::Repository;
 use log::info;
 use mdbook::renderer::RenderContext;
 use mdbook::MDBook;
@@ -12,7 +12,6 @@ use std::fs::File;
 use std::io::Write;
 use std::path::{Path, PathBuf};
 use tera::Context;
-use url::Url;
 use walkdir::WalkDir;
 
 /// A manifest entry for the generated EPUB
@@ -67,7 +66,7 @@ pub fn run(config: &Config) -> Result<Manifest, Error> {
         let repo_url = repo_config.repo_url.to_owned();
 
         let (mut repo_path, commit_sha, last_modified) =
-            clone_or_fetch_repo(repo_url.as_str(), working_dir)?;
+            git::clone_or_fetch_repo(repo_url.as_str(), working_dir)?;
 
         if let Some(repo_folder) = &repo_config.folder {
             repo_path = repo_path.join(repo_folder);
@@ -148,42 +147,6 @@ fn generate_epub(path: &Path, dest: &Path) -> Result<(Option<String>, PathBuf, u
     let title = md.config.book.title;
 
     Ok((title, output_path, epub_size))
-}
-
-/// Clones or fetches the repo at `entry.repo_url` inside `working_dir`.
-fn clone_or_fetch_repo(url: &str, working_dir: &Path) -> Result<(PathBuf, String, String), Error> {
-    let parsed_url = Url::parse(url)?;
-    // skip initial `/` in path
-    let mut dest = working_dir.join(&parsed_url.path()[1..]);
-
-    // :TRICKY: can't use \ as path separator here because of improper native path handling in some parts of libgit2
-    // see https://github.com/libgit2/libgit2/issues/3012
-    if cfg!(windows) {
-        dest = PathBuf::from(dest.to_str().unwrap().replace('\\', "/"));
-    }
-
-    let repo = if let Ok(repo) = Repository::open(&dest) {
-        repo.find_remote("origin").and_then(|mut remote| {
-            assert_eq!(
-                remote.url().unwrap(),
-                url,
-                "Remote url for origin and requested url do not match"
-            );
-            info!("Found {:?}. Fetching {}", dest, url);
-            remote.fetch(&["master"], None, None)
-        })?;
-        repo
-    } else {
-        // :TODO: shallow clone when supported by libgit2 (https://github.com/libgit2/libgit2/issues/3058)
-        info!("Cloning {} to {:?}", url, dest);
-        Repository::clone(url, &dest)?
-    };
-
-    let commit = repo.head()?.peel_to_commit()?;
-    let commit_sha = commit.id().to_string();
-    let last_modified = Utc.timestamp(commit.time().seconds(), 0).to_rfc3339();
-
-    Ok((dest, commit_sha, last_modified))
 }
 
 #[test]
